+++ xds/internal/xdsclient/client_new.go
@@ -27,7 +27,6 @@ import (
 	"time"
 
 	"google.golang.org/grpc/internal/cache"
-	"google.golang.org/grpc/internal/callbackserializer"
 	"google.golang.org/grpc/internal/grpcsync"
 	"google.golang.org/grpc/xds/internal/xdsclient/bootstrap"
 )
@@ -70,7 +69,7 @@ func newWithConfig(config *bootstrap.Config, watchExpiryTimeout time.Duration, i
 		done:               grpcsync.NewEvent(),
 		config:             config,
 		watchExpiryTimeout: watchExpiryTimeout,
-		serializer:         callbackserializer.New(ctx),
+		serializer:         grpcsync.NewCallbackSerializer(ctx),
 		serializerClose:    cancel,
 		resourceTypes:      newResourceTypeRegistry(),
 		authorities:        make(map[string]*authority),
