+++ xds/internal/xdsclient/clientimpl.go
@@ -23,7 +23,6 @@ import (
 	"time"
 
 	"google.golang.org/grpc/internal/cache"
-	"google.golang.org/grpc/internal/callbackserializer"
 	"google.golang.org/grpc/internal/grpclog"
 	"google.golang.org/grpc/internal/grpcsync"
 	"google.golang.org/grpc/xds/internal/xdsclient/bootstrap"
@@ -38,7 +37,7 @@ type clientImpl struct {
 	config             *bootstrap.Config
 	logger             *grpclog.PrefixLogger
 	watchExpiryTimeout time.Duration
-	serializer         *callbackserializer.CallbackSerializer
+	serializer         *grpcsync.CallbackSerializer
 	serializerClose    func()
 	resourceTypes      *resourceTypeRegistry
